Adds support for SameSite cookie attribute
